Fix .user.yml error with Fusion (#368)
## Summary

This fixes an issue when Fusion generates a `.user.yml` file in `<uuid>`
format rather than the `id: <uuid>` format that dbt Core uses. I think
this issue is being addressed in Fusion, but this should fix it in the
meantime. I reproed the issue with `task client` and ensured this fix
can handle the alternate format.

## Checklist
- [x] I have performed a self-review of my code
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes