Improved retry logic and post project selection screen (#383)
## Summary

This PR improves the project selection screen by adding retries to fetch
requests. Also, after a project is selected, a screen is shown to close
the browser tab. The browser tab should close automatically, but if it
doesn't, this screen should make it clear that the tab can be closed. I
tested this with `task client`. I commented out the `window.close()`
line to ensure the new screen works.

## Checklist
- [x] I have performed a self-review of my code
- [x] New and existing unit tests pass locally with my changes