Refactor proxied tools. Enable new search tools. (#443)
## Summary

This PR refactors the proxied tools (SQL tools) to include the new
search-based tools: `keyword_search` and `get_related_models`.

## What Changed

In a lot of places, "SQL tools" were renamed to "Proxied tools" to leave
room for the new search-based tools. `SQL` is still a separate toolset,
so there aren't breaking changes here. This change allows for the
proxying of two new tools.

## Why

To proxy the new search tools, I had to refactor the pieces of the code
that assumed SQL tools were the only kind that could be proxied.

## Checklist
- [x] I have performed a self-review of my code
- [x] I have made corresponding changes to the documentation (in
https://github.com/dbt-labs/docs.getdbt.com) if required -- Mention it
here
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes