Refactor discovery fetchers to share pagination (#451)
## Summary

This PR consolidates the pagination logic of the discovery tools. This
will make it easier to add new types of resources in the future.

## What Changed

Added `PaginatedResourceFetcher` to share pagination between the
different resource types.

## Why

Consolidates code and makes it easier to add resource types in the
future.

## Checklist
- [x] I have performed a self-review of my code
- [x] I have made corresponding changes to the documentation (in
https://github.com/dbt-labs/docs.getdbt.com) if required -- Mention it
here
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes