Close proxied_tools_manager for exceptions in listing proxied tools (#492)
## Summary

I have to do more research on anyio and why this was happening, but
first, I want to get this fix out. A combination of factors has led to
errors when trying to use Discovery & Semantic Layer tools.

First, we added a requirement that `x-dbt-user-id` header must not be
present when service token authentication is used. This change has led
to an increase in the number of unauthenticated requests to remote MCP.
Proxied tools rely on remote dbt MCP. Typically, if there are failures
loading the proxied tools, they are disabled, and this doesn't impact
the other tools. With this current bug, that is not the case.

This is because we migrated the Discovery & Semantic Layer tools to use
httpx rather than the requests library. For some reason that I still
have yet to understand, some underlying cancellation state of anyio
caused the Discovery & Semantic Layer httpx tool calls to fail when the
API call to list proxied tool calls failed. I will do more research.

This change reinstates the intended behavior, when listing proxied tools
fails, the other toolsets are not impacted.

## Checklist
- [x] I have performed a self-review of my code
- [x] I have made corresponding changes to the documentation (in
https://github.com/dbt-labs/docs.getdbt.com) if required -- Mention it
here
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] New and existing unit tests pass locally with my changes